# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package dbsqlc

import (
	"database/sql"
	"encoding/json"
	"time"

	"github.com/google/uuid"
)

type AiProposal struct {
	ID           uuid.UUID       `json:"id"`
	UserID       uuid.UUID       `json:"user_id"`
	ParentType   string          `json:"parent_type"`
	ParentID     uuid.UUID       `json:"parent_id"`
	MagicText    string          `json:"magic_text"`
	Summary      string          `json:"summary"`
	ProposalJson json.RawMessage `json:"proposal_json"`
	Status       string          `json:"status"`
	CreatedAt    time.Time       `json:"created_at"`
	AppliedAt    sql.NullTime    `json:"applied_at"`
}

type Epic struct {
	ID          uuid.UUID      `json:"id"`
	UserID      uuid.UUID      `json:"user_id"`
	Name        string         `json:"name"`
	Description sql.NullString `json:"description"`
	Color       sql.NullString `json:"color"`
	Position    int32          `json:"position"`
	CreatedAt   time.Time      `json:"created_at"`
	UpdatedAt   time.Time      `json:"updated_at"`
}

type Label struct {
	ID        uuid.UUID      `json:"id"`
	UserID    uuid.UUID      `json:"user_id"`
	Name      string         `json:"name"`
	Color     sql.NullString `json:"color"`
	CreatedAt time.Time      `json:"created_at"`
	UpdatedAt time.Time      `json:"updated_at"`
}

type Project struct {
	ID          uuid.UUID      `json:"id"`
	UserID      uuid.UUID      `json:"user_id"`
	EpicID      uuid.NullUUID  `json:"epic_id"`
	Name        string         `json:"name"`
	Description sql.NullString `json:"description"`
	Status      string         `json:"status"`
	StartDate   sql.NullTime   `json:"start_date"`
	TargetDate  sql.NullTime   `json:"target_date"`
	Position    int32          `json:"position"`
	CreatedAt   time.Time      `json:"created_at"`
	UpdatedAt   time.Time      `json:"updated_at"`
}

type ProjectNote struct {
	ID        uuid.UUID `json:"id"`
	UserID    uuid.UUID `json:"user_id"`
	ProjectID uuid.UUID `json:"project_id"`
	Body      string    `json:"body"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type Stage struct {
	ID          uuid.UUID      `json:"id"`
	UserID      uuid.UUID      `json:"user_id"`
	ProjectID   uuid.UUID      `json:"project_id"`
	Name        string         `json:"name"`
	Description sql.NullString `json:"description"`
	Status      string         `json:"status"`
	StartDate   sql.NullTime   `json:"start_date"`
	TargetDate  sql.NullTime   `json:"target_date"`
	Position    int32          `json:"position"`
	CreatedAt   time.Time      `json:"created_at"`
	UpdatedAt   time.Time      `json:"updated_at"`
}

type Todo struct {
	ID              uuid.UUID      `json:"id"`
	UserID          uuid.UUID      `json:"user_id"`
	ProjectID       uuid.NullUUID  `json:"project_id"`
	StageID         uuid.NullUUID  `json:"stage_id"`
	Title           string         `json:"title"`
	Description     sql.NullString `json:"description"`
	Priority        string         `json:"priority"`
	Status          string         `json:"status"`
	StartDate       sql.NullTime   `json:"start_date"`
	DueDate         sql.NullTime   `json:"due_date"`
	EstimatedEffort sql.NullString `json:"estimated_effort"`
	Position        int32          `json:"position"`
	NextAction      bool           `json:"next_action"`
	Milestone       bool           `json:"milestone"`
	Recurrence      string         `json:"recurrence"`
	CompletedAt     sql.NullTime   `json:"completed_at"`
	CreatedAt       time.Time      `json:"created_at"`
	UpdatedAt       time.Time      `json:"updated_at"`
}

type TodoDependency struct {
	TodoID          uuid.UUID `json:"todo_id"`
	DependsOnTodoID uuid.UUID `json:"depends_on_todo_id"`
	CreatedAt       time.Time `json:"created_at"`
}

type TodoLabel struct {
	TodoID  uuid.UUID `json:"todo_id"`
	LabelID uuid.UUID `json:"label_id"`
}

type User struct {
	ID            uuid.UUID `json:"id"`
	GoogleSubject string    `json:"google_subject"`
	Email         string    `json:"email"`
	Name          string    `json:"name"`
	Approved      bool      `json:"approved"`
	CreatedAt     time.Time `json:"created_at"`
	UpdatedAt     time.Time `json:"updated_at"`
}